Postman collection
Added a postman collection to make it easier to start using the API.
Also changed the JWT timeout to 7 days. This is much higher than normal but means the two variables adm_jwt and student_jwt specified for auth tokens in postman requests shouldn't expire all the time making it frustrating to test the API in postman.